refining taylor diagram and associated changes
